requirements.txt
discord.py
pygpt4all
